requirements.txt
Flask>=2.0.2
pymongo>=4.6.2
pyyaml>=6.0
